Add character type.

+++ src/main/java/org/indunet/fastproto/decoder/LongDecoder.java
@@ -5,7 +5,11 @@ import org.indunet.fastproto.annotation.type.LongType;
 import org.indunet.fastproto.exception.DecodeException;
 import org.indunet.fastproto.exception.DecodeException.DecodeError;
 
-
+/**
+ * @author Deng Ran
+ * @see TypeDecoder
+ * @since 1.0.0
+ */
 public class LongDecoder implements TypeDecoder<Long> {
     @Override
     public Long decode(DecodeContext context) {
@@ -23,22 +27,22 @@ public class LongDecoder implements TypeDecoder<Long> {
 
         if (endian == EndianPolicy.LITTLE) {
             value |= (datagram[byteOffset] & 0xFF);
-            value |= ((long) (datagram[byteOffset + 1] & 0xFFL) << 8);
-            value |= ((long) (datagram[byteOffset + 2] & 0xFFL) << 16);
-            value |= ((long) (datagram[byteOffset + 3] & 0xFFL) << 24);
-            value |= ((long) (datagram[byteOffset + 4] & 0xFFL) << 32);
-            value |= ((long) (datagram[byteOffset + 5] & 0xFFL) << 40);
-            value |= ((long) (datagram[byteOffset + 6] & 0xFFL) << 48);
-            value |= ((long) (datagram[byteOffset + 7] & 0xFFL) << 56);
+            value |= ((datagram[byteOffset + 1] & 0xFFL) << 8);
+            value |= ((datagram[byteOffset + 2] & 0xFFL) << 16);
+            value |= ((datagram[byteOffset + 3] & 0xFFL) << 24);
+            value |= ((datagram[byteOffset + 4] & 0xFFL) << 32);
+            value |= ((datagram[byteOffset + 5] & 0xFFL) << 40);
+            value |= ((datagram[byteOffset + 6] & 0xFFL) << 48);
+            value |= ((datagram[byteOffset + 7] & 0xFFL) << 56);
         } else if (endian == EndianPolicy.BIG) {
             value |= (datagram[byteOffset + 7] & 0xFF);
-            value |= ((long) (datagram[byteOffset + 6] & 0xFFL) << 8);
-            value |= ((long) (datagram[byteOffset + 5] & 0xFFL) << 16);
-            value |= ((long) (datagram[byteOffset + 4] & 0xFFL) << 24);
-            value |= ((long) (datagram[byteOffset + 3] & 0xFFL) << 32);
-            value |= ((long) (datagram[byteOffset + 2] & 0xFFL) << 40);
-            value |= ((long) (datagram[byteOffset + 1] & 0xFFL) << 48);
-            value |= ((long) (datagram[byteOffset] & 0xFFL) << 56);
+            value |= ((datagram[byteOffset + 6] & 0xFFL) << 8);
+            value |= ((datagram[byteOffset + 5] & 0xFFL) << 16);
+            value |= ((datagram[byteOffset + 4] & 0xFFL) << 24);
+            value |= ((datagram[byteOffset + 3] & 0xFFL) << 32);
+            value |= ((datagram[byteOffset + 2] & 0xFFL) << 40);
+            value |= ((datagram[byteOffset + 1] & 0xFFL) << 48);
+            value |= ((datagram[byteOffset] & 0xFFL) << 56);
         }
 
         return value;
